Introduce DownloaderInterface and move FileDownloader to Downloader namespace
