+++ src/Command/AbstractUpdateCommand.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Phpcq\Runner\Command;
 
-use Phpcq\Runner\FileDownloader;
+use Phpcq\Runner\Downloader\FileDownloader;
 use Phpcq\GnuPG\Downloader\KeyDownloader;
 use Phpcq\GnuPG\GnuPGFactory;
 use Phpcq\GnuPG\Signature\AlwaysStrategy;
